+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next]
 
 * feat: Added `NavigationPane.toggleable` ([#973](https://github.com/bdlukaa/fluent_ui/issues/973))
+* feat: `NumberBox.onTextChange` ([#972](https://github.com/bdlukaa/fluent_ui/issues/972))
 
 ## 4.8.0